Add type archive for 2.10
