+++ open-metadata-resources/open-metadata-archives/open-metadata-types/src/main/java/org/odpi/openmetadata/opentypes/OpenMetadataTypesArchive.java
@@ -37,7 +37,7 @@ public class OpenMetadataTypesArchive
     private static final String                  archiveName        = "Open Metadata Types";
     private static final String                  archiveDescription = "Standard types for open metadata repositories.";
     private static final OpenMetadataArchiveType archiveType        = OpenMetadataArchiveType.CONTENT_PACK;
-    private static final String                  archiveVersion     = "2.9";
+    private static final String                  archiveVersion     = "2.10";
     private static final String                  originatorName     = "Egeria";
     private static final String                  originatorLicense  = "Apache 2.0";
     private static final Date                    creationDate       = new Date(1588261366992L);
@@ -145,7 +145,7 @@ public class OpenMetadataTypesArchive
      */
     public void getOriginalTypes()
     {
-        OpenMetadataTypesArchive2_8 previousTypes = new OpenMetadataTypesArchive2_8(archiveBuilder);
+        OpenMetadataTypesArchive2_9 previousTypes = new OpenMetadataTypesArchive2_9(archiveBuilder);
 
         /*
          * Pull the types from previous releases.
@@ -155,346 +155,7 @@ public class OpenMetadataTypesArchive
         /*
          * Calls for new and changed types go here
          */
-        update0056AssetManagers();
-        update0137ToDos();
-        update0450GovernanceRollout();
-        update0217Ports();
-        update0534RelationalSchema();
-        add0335PrimaryCategoryClassification();
-    }
 
-
-
-
-    /*
-     * -------------------------------------------------------------------------------------------------------
-     */
-
-    /**
-     * Add UserProfileManager, UserAccessDirectory and MasterDataManager classification for Referenceables.
-     */
-    private void update0056AssetManagers()
-    {
-        this.archiveBuilder.addClassificationDef(getUserProfileManagerClassification());
-        this.archiveBuilder.addClassificationDef(getUserAccessDirectoryClassification());
-        this.archiveBuilder.addClassificationDef(getMasterDataManagerClassification());
-    }
-
-    private ClassificationDef getUserProfileManagerClassification()
-    {
-        final String guid            = "53ef4062-9e0a-4892-9824-8d51d4ad59d3";
-        final String name            = "UserProfileManager";
-        final String description     = "A system that sores descriptions of individuals and their roles/interests in an organization.";
-        final String descriptionGUID = null;
-
-        final String linkedToEntity = "Referenceable";
-
-        return archiveHelper.getClassificationDef(guid,
-                                                  name,
-                                                  null,
-                                                  description,
-                                                  descriptionGUID,
-                                                  this.archiveBuilder.getEntityDef(linkedToEntity),
-                                                  false);
-    }
-
-    private ClassificationDef getUserAccessDirectoryClassification()
-    {
-        final String guid            = "29c98cf7-32b3-47d2-a411-48c1c9967e6d";
-        final String name            = "UserAccessDirectory";
-        final String description     = "A system that stores the access rights and groups for users (people and automated processes).";
-        final String descriptionGUID = null;
-
-        final String linkedToEntity = "Referenceable";
-
-        return archiveHelper.getClassificationDef(guid,
-                                                  name,
-                                                  null,
-                                                  description,
-                                                  descriptionGUID,
-                                                  this.archiveBuilder.getEntityDef(linkedToEntity),
-                                                  false);
-    }
-
-    private ClassificationDef getMasterDataManagerClassification()
-    {
-        final String guid            = "5bdad12e-57e7-4ff9-b7be-5d869e77d30b";
-        final String name            = "MasterDataManager";
-        final String description     = "A system that manages the consolidation and reconciliation of master data - typically people, organizations, products and accounts.";
-        final String descriptionGUID = null;
-
-        final String linkedToEntity = "Referenceable";
-
-        return archiveHelper.getClassificationDef(guid,
-                                                  name,
-                                                  null,
-                                                  description,
-                                                  descriptionGUID,
-                                                  this.archiveBuilder.getEntityDef(linkedToEntity),
-                                                  false);
-    }
-
-
-
-    /*
-     * -------------------------------------------------------------------------------------------------------
-     */
-
-    /**
-     * Add ActionTarget relationship for ToDos.
-     */
-    private void update0137ToDos()
-    {
-        this.archiveBuilder.addRelationshipDef(getActionTargetRelationship());
-    }
-
-
-    private RelationshipDef getActionTargetRelationship()
-    {
-        final String guid            = "207e2594-e3e4-4be8-a12c-4c401656e241";
-        final String name            = "ActionTarget";
-        final String description     = "Associates a To Do with one or more elements to work on.";
-        final String descriptionGUID = null;
-
-        final ClassificationPropagationRule classificationPropagationRule = ClassificationPropagationRule.NONE;
-
-        RelationshipDef relationshipDef = archiveHelper.getBasicRelationshipDef(guid,
-                                                                                name,
-                                                                                null,
-                                                                                description,
-                                                                                descriptionGUID,
-                                                                                classificationPropagationRule);
-
-        RelationshipEndDef relationshipEndDef;
-
-        /*
-         * Set up end 1.
-         */
-        final String                     end1EntityType               = "ToDo";
-        final String                     end1AttributeName            = "identifiedToDoActions";
-        final String                     end1AttributeDescription     = "Actions that have been identified for this element.";
-        final String                     end1AttributeDescriptionGUID = null;
-        final RelationshipEndCardinality end1Cardinality              = RelationshipEndCardinality.ANY_NUMBER;
-
-        relationshipEndDef = archiveHelper.getRelationshipEndDef(this.archiveBuilder.getEntityDef(end1EntityType),
-                                                                 end1AttributeName,
-                                                                 end1AttributeDescription,
-                                                                 end1AttributeDescriptionGUID,
-                                                                 end1Cardinality);
-        relationshipDef.setEndDef1(relationshipEndDef);
-
-
-        /*
-         * Set up end 2.
-         */
-        final String                     end2EntityType               = "Referenceable";
-        final String                     end2AttributeName            = "elementsToWorkOn";
-        final String                     end2AttributeDescription     = "Elements that will be updated or used to complete the action.";
-        final String                     end2AttributeDescriptionGUID = null;
-        final RelationshipEndCardinality end2Cardinality              = RelationshipEndCardinality.ANY_NUMBER;
-
-        relationshipEndDef = archiveHelper.getRelationshipEndDef(this.archiveBuilder.getEntityDef(end2EntityType),
-                                                                 end2AttributeName,
-                                                                 end2AttributeDescription,
-                                                                 end2AttributeDescriptionGUID,
-                                                                 end2Cardinality);
-        relationshipDef.setEndDef2(relationshipEndDef);
-
-        return relationshipDef;
-    }
-
-
-
-    /*
-     * -------------------------------------------------------------------------------------------------------
-     */
-
-    /**
-     * Add GovernanceMeasurements classification for Referenceables.
-     */
-    private void update0450GovernanceRollout()
-    {
-        this.archiveBuilder.addClassificationDef(getGovernanceMeasurementClassification());
-    }
-
-    private ClassificationDef getGovernanceMeasurementClassification()
-    {
-        final String guid            = "9d99d962-0214-49ba-83f7-c9b1f9f5bed4";
-        final String name            = "GovernanceMeasurements";
-        final String description     = "A set of measurements on the performance and use of the connected resource.";
-        final String descriptionGUID = null;
-
-        final String linkedToEntity = "Referenceable";
-
-        ClassificationDef classificationDef = archiveHelper.getClassificationDef(guid,
-                                                                                 name,
-                                                                                 null,
-                                                                                 description,
-                                                                                 descriptionGUID,
-                                                                                 this.archiveBuilder.getEntityDef(linkedToEntity),
-                                                                                 false);
-
-        /*
-         * Build the attributes
-         */
-        List<TypeDefAttribute> properties = new ArrayList<>();
-        TypeDefAttribute       property;
-
-        final String attribute1Name            = "measurementCounts";
-        final String attribute1Description     = "A set of metric name to current count value pairs.";
-        final String attribute1DescriptionGUID = null;
-        final String attribute2Name            = "measurementValues";
-        final String attribute2Description     = "A set of metric name to current value pairs.";
-        final String attribute2DescriptionGUID = null;
-        final String attribute3Name            = "measurementFlags";
-        final String attribute3Description     = "A set of metric name to current boolean value pairs.";
-        final String attribute3DescriptionGUID = null;
-
-        property = archiveHelper.getMapStringIntTypeDefAttribute(attribute1Name,
-                                                                 attribute1Description,
-                                                                 attribute1DescriptionGUID);
-        properties.add(property);
-        property = archiveHelper.getMapStringStringTypeDefAttribute(attribute2Name,
-                                                                    attribute2Description,
-                                                                    attribute2DescriptionGUID);
-        properties.add(property);
-        property = archiveHelper.getMapStringBooleanTypeDefAttribute(attribute3Name,
-                                                                     attribute3Description,
-                                                                     attribute3DescriptionGUID);
-        properties.add(property);
-
-        classificationDef.setPropertiesDefinition(properties);
-
-        return classificationDef;
-    }
-
-
-    /*
-     * -------------------------------------------------------------------------------------------------------
-     */
-
-    /**
-     * Add filterExpression attribute to Port.
-     */
-    private void update0217Ports()
-    {
-        this.archiveBuilder.addTypeDefPatch(updatePortEntity());
-    }
-
-
-    /**
-     * Add filterExpression attribute to Port.
-     * @return the type def patch
-     */
-    private TypeDefPatch updatePortEntity()
-    {
-        /*
-         * Create the Patch
-         */
-        final String typeName = "Port";
-
-        TypeDefPatch typeDefPatch = archiveBuilder.getPatchForType(typeName);
-
-        typeDefPatch.setUpdatedBy(originatorName);
-        typeDefPatch.setUpdateTime(creationDate);
-
-        /*
-         * Build the attributes
-         */
-        List<TypeDefAttribute> properties = new ArrayList<>();
-        TypeDefAttribute       property;
-
-        final String attribute1Name            = "filterExpression";
-        final String attribute1Description     = "Expression used to filter data values passing through port.";
-        final String attribute1DescriptionGUID = null;
-
-        property = archiveHelper.getStringTypeDefAttribute(attribute1Name,
-                                                           attribute1Description,
-                                                           attribute1DescriptionGUID);
-        properties.add(property);
-
-        typeDefPatch.setPropertyDefinitions(properties);
-
-        return typeDefPatch;
-    }
-
-
-
-    /*
-     * -------------------------------------------------------------------------------------------------------
-     */
-
-
-    /**
-     * The RelationalColumnType only allows for a column to be primitive. - could be a literal, enum or external.
-     */
-    private void update0534RelationalSchema()
-    {
-        this.archiveBuilder.addTypeDefPatch(deprecateRelationalColumnType());
-    }
-
-    private TypeDefPatch deprecateRelationalColumnType()
-    {
-        final String typeName = "RelationalColumnType";
-
-        TypeDefPatch typeDefPatch = archiveBuilder.getPatchForType(typeName);
-
-        typeDefPatch.setUpdatedBy(originatorName);
-        typeDefPatch.setUpdateTime(creationDate);
-        typeDefPatch.setTypeDefStatus(TypeDefStatus.DEPRECATED_TYPEDEF);
-
-        return typeDefPatch;
-    }
-
-
-
-    /*
-     * -------------------------------------------------------------------------------------------------------
-     */
-
-    /**
-     * 0335  Add the Primary Category Classification
-     */
-    private void add0335PrimaryCategoryClassification()
-    {
-        this.archiveBuilder.addClassificationDef(addPrimaryCategoryClassification());
-    }
-
-    private ClassificationDef addPrimaryCategoryClassification()
-    {
-        final String guid            = "3a6c4ba7-3cc5-48cd-8952-bwra92da016d";
-        final String name            = "PrimaryCategory";
-        final String description     = "Defines a category as being the base category of a glossary term";
-        final String descriptionGUID = null;
-
-        final String linkedToEntity = "GlossaryTerm";
-
-        ClassificationDef classificationDef = archiveHelper.getClassificationDef(guid,
-                                                                                 name,
-                                                                                 null,
-                                                                                 description,
-                                                                                 descriptionGUID,
-                                                                                 this.archiveBuilder.getEntityDef(linkedToEntity),
-                                                                                 true);
-
-        /*
-         * Build the attributes
-         */
-        List<TypeDefAttribute> properties = new ArrayList<>();
-        TypeDefAttribute       property;
-
-        final String attribute1Name            = "categoryQualifiedName";
-        final String attribute1Description     = "The qualified name of the primary category of a GlossaryTerm.";
-        final String attribute1DescriptionGUID = null;
-
-        property = archiveHelper.getStringTypeDefAttribute(attribute1Name,
-                                                           attribute1Description,
-                                                           attribute1DescriptionGUID);
-        properties.add(property);
-
-        classificationDef.setPropertiesDefinition(properties);
-
-        return classificationDef;
     }
 }
